Adjust synctask package
