+++ sharding-scaling/sharding-scaling-core/src/main/java/org/apache/shardingsphere/shardingscaling/core/controller/task/ReportCallback.java
@@ -15,19 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.shardingscaling.core.synctask;
+package org.apache.shardingsphere.shardingscaling.core.controller.task;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.apache.shardingsphere.shardingscaling.core.controller.SyncProgress;
+import org.apache.shardingsphere.shardingscaling.core.execute.Event;
 
-@Getter
-@AllArgsConstructor
-public class HistoryDataSyncTaskProgress implements SyncProgress {
+/**
+ * Report callback.
+ *
+ * @author avalon566
+ */
+public interface ReportCallback {
 
-    private final String id;
-
-    private final long estimatedRows;
-
-    private final long syncedRows;
+    /**
+     * process report event.
+     *
+     * @param event report event
+     */
+    void onProcess(Event event);
 }
